# schemas/ripd/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RIP Agent Server",
    "description": "Agent session control plane (HTTP/SSE).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "operationId": "list_tasks",
        "responses": {
          "200": {
            "description": "List tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskStatusResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSpawnPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid task request"
          }
        }
      }
    },
...
    "/tasks/{id}/events": {
      "get": {
        "operationId": "stream_task_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of task event frames"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
...
    "/threads/{id}/branch": {
      "post": {
        "operationId": "thread_branch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Parent thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadBranchPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadBranchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid branch request"
          },
          "404": {
            "description": "Thread or branch point not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-auto": {
      "post": {
        "operationId": "thread_compaction_auto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionAutoV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Auto-compaction no-op / dry-run result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoV1Response"
                }
              }
            }
          },
          "202": {
            "description": "Auto-compaction job spawned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid auto-compaction request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-auto-schedule": {
      "post": {
        "operationId": "thread_compaction_auto_schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionAutoScheduleV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scheduler no-op / dry-run / skipped result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoScheduleV1Response"
                }
              }
            }
          },
          "202": {
            "description": "Scheduler started a job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoScheduleV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid schedule request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-checkpoint": {
      "post": {
        "operationId": "thread_compaction_checkpoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCompactionCheckpointPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Compaction checkpoint created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadCompactionCheckpointResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid checkpoint request"
          },
          "404": {
            "description": "Thread or cut point not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-cut-points": {
      "post": {
        "operationId": "thread_compaction_cut_points",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionCutPointsV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Computed cut points",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionCutPointsV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cut point request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-status": {
      "post": {
        "operationId": "thread_compaction_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionStatusV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Compaction status projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionStatusV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
...
    "/threads/{id}/handoff": {
      "post": {
        "operationId": "thread_handoff",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadHandoffPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Handoff thread created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadHandoffResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid handoff request"
          },
          "404": {
            "description": "Thread or handoff point not found"
          }
        }
      }
    },
    "/threads/{id}/messages": {
      "post": {
        "operationId": "thread_post_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadPostMessagePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Message accepted and run started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadPostMessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiToolTaskExecutionMode": {
        "type": "string",
        "enum": [
          "pipes",
          "pty"
        ]
      },
...
      "CompactionAutoScheduleV1Request": {
        "type": "object",
        "required": [
          "actor_id",
          "origin"
        ],
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "block_on_inflight": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "dry_run": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "execute": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "max_new_checkpoints": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "origin": {
            "type": "string"
          },
          "stride_messages": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "CompactionAutoScheduleV1Response": {
        "type": "object",
        "required": [
          "thread_id",
          "policy_id",
          "decision",
          "execute",
          "stride_messages",
          "max_new_checkpoints",
          "block_on_inflight",
          "message_count",
          "cut_rule_id",
          "planned",
          "result"
        ],
        "properties": {
          "block_on_inflight": {
            "type": "boolean"
          },
          "cut_rule_id": {
            "type": "string"
          },
          "decision": {
            "type": "string"
          },
          "decision_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "execute": {
            "type": "boolean"
          },
          "job_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_new_checkpoints": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "message_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "planned": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompactionPlannedCutPointV1"
            }
          },
          "policy_id": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompactionAutoResultCheckpointV1"
            }
          },
          "stride_messages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
      "CompactionAutoV1Request": {
        "type": "object",
        "required": [
          "actor_id",
          "origin"
        ],
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "dry_run": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "max_new_checkpoints": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "origin": {
            "type": "string"
          },
          "stride_messages": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "CompactionAutoV1Response": {
        "type": "object",
        "required": [
          "thread_id",
          "status",
          "stride_messages",
          "message_count",
          "cut_rule_id",
          "planned",
          "result"
        ],
        "properties": {
          "cut_rule_id": {
            "type": "string"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "message_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "planned": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompactionPlannedCutPointV1"
            }
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompactionAutoResultCheckpointV1"
            }
          },
          "status": {
            "type": "string"
          },
          "stride_messages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
...
      "CompactionCutPointsV1Request": {
        "type": "object",
        "properties": {
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "stride_messages": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "TaskCreated": {
        "type": "object",
        "required": [
          "task_id"
        ],
        "properties": {
          "task_id": {
            "type": "string"
          }
        }
      },
...
      "TaskSpawnPayload": {
        "type": "object",
        "required": [
          "tool"
        ],
        "properties": {
          "args": {},
          "execution_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiToolTaskExecutionMode"
              }
            ]
          },
          "origin_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "tool": {
            "type": "string"
          }
        }
      },
      "TaskStatusResponse": {
        "type": "object",
        "required": [
          "task_id",
          "status",
          "tool",
          "execution_mode"
        ],
        "properties": {
          "artifacts": {},
          "ended_at_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ApiToolTaskExecutionMode"
          },
          "exit_code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "started_at_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "status": {
            "$ref": "#/components/schemas/ApiToolTaskStatus"
          },
          "task_id": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "tool": {
            "type": "string"
          }
        }
      },
...
      "ThreadBranchPayload": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "origin": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ThreadBranchResponse": {
        "type": "object",
        "required": [
          "thread_id",
          "parent_thread_id",
          "parent_seq"
        ],
        "properties": {
          "parent_message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "parent_thread_id": {
            "type": "string"
          },
          "thread_id": {
            "type": "string"
          }
        }
      },
      "ThreadCompactionCheckpointPayload": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "origin": {
            "type": [
              "string",
              "null"
            ]
          },
          "stride_messages": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "summary_artifact_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary_markdown": {
            "type": [
              "string",
              "null"
            ]
          },
          "to_message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "to_seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    }
  }
}